Finished boss fight, starting final endgame procedure

+++ BlankProject/Assets/Scripts/IDamage.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IDamage 
 {
-    void takeDamage(int amount);
+    void takeDamage(float amount);
 
-    void criticalHit(int amount);
+    void criticalHit(float amount);
 }
