+++ BlankProject/Assets/Scripts/playerDamage.cs
@@ -129,7 +129,7 @@ public class playerDamage : MonoBehaviour, IDamage
         return false;
     }
 
-    public void takeDamage(int amount)
+    public void takeDamage(float amount)
     {
         HP -= amount;
         isTakingDamage = true;
@@ -168,7 +168,7 @@ public class playerDamage : MonoBehaviour, IDamage
         regenCoroutine = null;
     }
 
-    public void criticalHit(int amount)
+    public void criticalHit(float amount)
     {
 
     }
